test: extract the writer so the invariant is a file boundary
The sessions folder was over the per-folder test budget, and my new file was the
reason: it was named after a concept rather than a unit, which is the pattern the
budget rule exists to catch.

Extracting `settleLifecycleWriter.ts` resolves that properly instead of adding an
exception. The test becomes a plain colocated `settleLifecycleWriter.test.ts`
next to the module it covers, matching the four files already in the folder, and
`sessionService.ts` drops 1945 → 1817 lines.

It also makes the invariant stronger. The source scan no longer allowlists a
region delimited by string offsets inside a 1800-line service — it allowlists a
FILE, which no future reordering can silently widen. Verified non-vacuous:
re-adding `settled_at = null` to a raw statement in `sessionService.ts` fails the
test and names the offending file.

Docs: the writer is in the terminals source map with the exact scope of what the
revision detects, and §3a records the file-boundary allowlist.
